Restore parser.rs with CSV detection and simple Nexpose routing
