+++ src/parser.rs
@@ -1,10 +1,12 @@
-//! Utilities for parsing Nessus XML reports into in-memory models.
+//! Utilities for parsing vulnerability scan reports into in-memory models.
+
+mod simple_nexpose;
 
 use std::collections::BTreeSet;
 use std::path::Path;
 
-use quick_xml::Reader;
 use quick_xml::events::Event;
+use quick_xml::Reader;
 use tracing::{debug, info};
 
 use crate::models::{Host, Item, Plugin};
@@ -18,8 +20,24 @@ pub struct NessusReport {
     pub plugins: Vec<Plugin>,
 }
 
-/// Validate and parse a Nessus XML file into ORM models.
+/// Detect file type and parse accordingly.
 pub fn parse_file(path: &Path) -> Result<NessusReport, crate::error::Error> {
+    match path
+        .extension()
+        .and_then(|e| e.to_str())
+        .map(|s| s.to_lowercase())
+        .as_deref()
+    {
+        Some("csv") => {
+            let report = simple_nexpose::parse_file(path)?;
+            Ok(report.into())
+        }
+        _ => parse_nessus(path),
+    }
+}
+
+/// Validate and parse a Nessus XML file into ORM models.
+fn parse_nessus(path: &Path) -> Result<NessusReport, crate::error::Error> {
     info!("Parsing file: {}", path.display());
 
     let mut reader = Reader::from_file(path)?;
@@ -182,6 +200,15 @@ mod tests {
         assert_eq!(report.hosts[0].ip.as_deref(), Some("192.168.0.1"));
         assert_eq!(report.plugins.len(), 1);
     }
+
+    #[test]
+    fn routes_csv_to_simple_parser() {
+        let path = std::path::Path::new("tests/fixtures/sample_nexpose.csv");
+        let report = parse_file(path).expect("parse csv");
+        assert_eq!(report.hosts.len(), 2);
+        assert_eq!(report.items.len(), 3);
+        assert_eq!(report.plugins.len(), 2);
+    }
 }
 
 fn empty_host() -> Host {
@@ -233,6 +260,7 @@ fn empty_item() -> Item {
         engagement_id: None,
     }
 }
+
 
 fn empty_plugin() -> Plugin {
     Plugin {
